有效算力、算力增速 修改数据类型为decimal

+++ miner/miner-server/src/main/java/com/mei/hui/miner/entity/SysAggPowerDaily.java
@@ -2,11 +2,9 @@ package com.mei.hui.miner.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.Date;
 
 /**
  * 算力按天聚合对象 sys_agg_power_daily
@@ -27,10 +25,10 @@ public class SysAggPowerDaily{
     private String date;
 
     /** 有效算力 */
-    private Long powerAvailable;
+    private BigDecimal powerAvailable;
 
     /** 算力增长 */
-    private Long powerIncrease;
+    private BigDecimal powerIncrease;
 
     /** 创建时间 */
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
